moved MultiLocation-converts to runtime-common
