+++ runtime/common/Cargo.toml
@@ -9,26 +9,37 @@ publish = false
 
 
 [dependencies]
+codec = { package = "parity-scale-codec", workspace = true, default-features = false }
+cumulus-primitives-core = { workspace = true, default-features = false }
+cumulus-pallet-parachain-system = { workspace = true, default-features = false }
 frame-support = { workspace = true, default-features = false }
 frame-system = { workspace = true, default-features = false }
 orml-tokens = { workspace = true, default-features = false }
 orml-traits = { workspace = true, default-features = false }
 pallet-transaction-payment = { workspace = true, default-features = false }
 peaq-primitives-xcm = { path = "../../primitives/xcm", default-features = false }
+sp-core = { workspace = true, default-features = false }
 sp-runtime = { workspace = true, default-features = false }
 sp-std = { workspace = true, default-features = false }
+xcm = { workspace = true, default-features = false }
 zenlink-protocol = { workspace = true, default-features = false }
 
 
 [features]
 default = ["std"]
 std = [
-	"sp-runtime/std",
-	"pallet-transaction-payment/std",
-	"peaq-primitives-xcm/std",
+	"codec/std",
+	"cumulus-primitives-core/std",
+	"cumulus-pallet-parachain-system/std",
 	"frame-support/std",
+	"frame-system/std",
 	"orml-tokens/std",
 	"orml-traits/std",
-	"frame-system/std",
+	"pallet-transaction-payment/std",
+	"peaq-primitives-xcm/std",
+	"sp-core/std",
+	"sp-runtime/std",
+	"sp-std/std",
+	"xcm/std",
 	"zenlink-protocol/std",
 ]
