+++ runtime/common/src/lib.rs
@@ -2,26 +2,35 @@
 #![recursion_limit = "256"]
 
 
+use codec::{Decode, Encode};
+use cumulus_primitives_core::ParaId;
 use frame_support::{
 	parameter_types,
 	pallet_prelude::{DispatchError, DispatchResult},
 	traits::Get,
 };
+use frame_system::Config as SystemT;
 use orml_traits::{
 	MultiCurrency,
 	currency::MutationHooks,
 };
-use sp_std::marker::PhantomData;
-use sp_runtime::Perbill;
+use cumulus_pallet_parachain_system::Config as ParachainSystemT;
+use sp_core::bounded::BoundedVec;
+use sp_std::{
+	marker::PhantomData, vec::Vec,
+};
+use sp_runtime::{
+	Perbill, RuntimeString,
+	traits::Convert,
+};
 use zenlink_protocol::{
 	AssetBalance, AssetId as ZenlinkAssetId, LocalAssetHandler,
 };
+use xcm::latest::prelude::*;
+use peaq_primitives_xcm::{
+	AccountId, Balance, CurrencyId, TokenSymbol, currency::parachain,
+};
 
-
-// /// Balance of an account.
-// pub type Balance = peaq_primitives_xcm::Balance;
-
-use peaq_primitives_xcm::{Balance, CurrencyId};
 
 // Contracts price units.
 pub const TOKEN_DECIMALS: u32 = 18;
@@ -145,4 +154,158 @@ where
 
 		Ok(amount)
 	}
+}
+
+
+/// A MultiLocation-AccountId converter for XCM, Zenlink-Protocol and similar stuff.
+pub struct AccountIdToMultiLocation;
+
+impl Convert<AccountId, MultiLocation> for AccountIdToMultiLocation {
+	fn convert(account: AccountId) -> MultiLocation {
+		X1(AccountId32 { network: None, id: account.into() }).into()
+	}
+}
+
+
+/// A MultiLocation-CurrencyId converter for XCM, Zenlink-Protocol and similar stuff.
+pub struct CurrencyIdConvert<T>(PhantomData<T>) where T: SystemT + ParachainSystemT;
+
+impl<T> Convert<CurrencyId, Option<MultiLocation>> for CurrencyIdConvert<T>
+where
+	T: SystemT + ParachainSystemT,
+{
+	fn convert(id: CurrencyId) -> Option<MultiLocation> {
+		use CurrencyId::Token;
+		use TokenSymbol::*;
+		// use parachain_info::pallet::
+
+		match id {
+			Token(DOT) | Token(KSM) | Token(ROC) =>
+				Some(MultiLocation::parent()),
+			Token(PEAQ) =>
+				native_currency_location(
+					<T as ParachainSystemT>::SelfParaId::get().into(),
+					id.encode()
+				),
+			Token(ACA) =>
+				native_currency_location(
+					parachain::acala::ID,
+					parachain::acala::ACA_KEY.to_vec(),
+				),
+			Token(BNC) =>
+				native_currency_location(
+					parachain::bifrost::ID,
+					parachain::bifrost::BNC_KEY.to_vec(),
+				),
+			_ => None,
+		}
+	}
+}
+
+impl<T> Convert<MultiLocation, Option<CurrencyId>> for CurrencyIdConvert<T>
+where
+	T: SystemT + ParachainSystemT,
+{
+	fn convert(location: MultiLocation) -> Option<CurrencyId> {
+		use CurrencyId::Token;
+		use TokenSymbol::*;
+		use RuntimeString::Borrowed as RsBorrowed;
+
+		match location {
+			MultiLocation {
+				parents: 1,
+				interior: Here,
+			} => {
+				let version = <T as SystemT>::Version::get();
+				match version.spec_name {
+					RsBorrowed("peaq-node-dev") => Some(Token(DOT)),
+					RsBorrowed("peaq-node-agung") => Some(Token(ROC)),
+					RsBorrowed("peaq-node-krest") => Some(Token(KSM)),
+					RsBorrowed("peaq-node") => Some(Token(DOT)),
+					_ => None,
+				}
+				// return Some(Token(DOT))
+			},
+			MultiLocation {
+				parents: 1,
+				interior: X2(Parachain(id), GeneralKey{ data, length })
+			} => {
+				let key = &data[..data.len().min(length as usize)];
+				match id {
+					parachain::acala::ID => {
+						match key {
+							parachain::acala::ACA_KEY => Some(Token(ACA)),
+							_ => None,
+						}
+					},
+					parachain::bifrost::ID => {
+						match key {
+							parachain::bifrost::BNC_KEY => Some(Token(BNC)),
+							_ => None,
+						}
+					},
+					_ => {
+						if ParaId::from(id) == <T as ParachainSystemT>::SelfParaId::get() {
+							if let Ok(currency_id) = CurrencyId::decode(&mut &*key) {
+								match currency_id {
+									Token(PEAQ) => Some(currency_id),
+									_ => None,
+								}
+							} else {
+								None
+							}
+						} else {
+							None
+						}
+					}
+				}
+			},
+			MultiLocation {
+				parents: 0,
+				interior: X1(GeneralKey { data, length })
+			} => {
+				let key = &data[..data.len().min(length as usize)];
+				// decode the general key
+				if let Ok(currency_id) = CurrencyId::decode(&mut &*key) {
+					match currency_id {
+						Token(PEAQ) => Some(currency_id),
+						_ => None,
+					}
+				} else {
+					None
+				}
+			},
+			_ => None,
+		}
+	}
+}
+
+impl<T> Convert<MultiAsset, Option<CurrencyId>> for CurrencyIdConvert<T>
+where
+	T: SystemT + ParachainSystemT,
+{
+	fn convert(asset: MultiAsset) -> Option<CurrencyId> {
+		if let MultiAsset { id: Concrete(location), .. } = asset {
+			Self::convert(location)
+		} else {
+			None
+		}
+	}
+}
+
+pub fn native_currency_location(para_id: u32, key: Vec<u8>) -> Option<MultiLocation> {
+	Some(MultiLocation::new(
+		1,
+		X2(
+			Parachain(para_id),
+			Junction::from(BoundedVec::try_from(key).ok()?)
+		),
+	))
+}
+
+pub fn local_currency_location(key: CurrencyId) -> Option<MultiLocation> {
+	Some(MultiLocation::new(
+		0,
+		X1(Junction::from(BoundedVec::try_from(key.encode()).ok()?)),
+	))
 }
